Added a check for file existence before saving
Added check for file existence before saving

+++ src/seedu/addressbook/storage/StorageFile.java
@@ -88,6 +88,10 @@ public class StorageFile {
         /* Note: Note the 'try with resource' statement below.
          * More info: https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html
          */
+        if (!path.toFile().exists() || !path.toFile().isFile()) {
+            throw new StorageOperationException("File is missing: " + path);
+        }
+        
         try (final Writer fileWriter =
                      new BufferedWriter(new FileWriter(path.toFile()))) {
